# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ParkWell - Parking Attendant System API",
    "description": "API documentation for the ParkWell system, for administrators and parking attendants.",
    "version": "1.1.0",
    "contact": {
      "name": "ParkWell Support",
      "email": "support@parkwell.example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development Server (v1)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/register-staff": {
      "post": {
        "tags": [
          "Authentication (Staff)"
        ],
        "summary": "Register a new staff member",
        "operationId": "registerStaff",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Staff account created. Verification email sent.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Specific error message"
          }
        }
      },
...
      "RoleNameEnum": {
        "type": "string",
        "enum": [
          "PARKING_ATTENDANT",
          "ADMIN"
        ],
        "example": "PARKING_ATTENDANT"
      },
...
      "StaffRegisterRequest": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "staff@parkwell.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Str0ngP@ss!"
          },
          "roleName": {
            "$ref": "#/components/schemas/RoleNameEnum",
            "description": "Optional, for Admin creating user. Defaults to PARKING_ATTENDANT."
          }
        }
      },
...
    },
...
    "responses": {
...
      "BadRequestError": {
        "description": "Bad Request. Invalid input or validation error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}